+++ readium/shared/src/main/java/org/readium/r2/shared/publication/Locator.kt
@@ -68,7 +68,10 @@ public data class Locator(
         val otherLocations: @WriteWith<JSONParceler> Map<String, Any> = emptyMap(),
     ) : JSONable, Parcelable {
 
-        override fun toJSON(): JSONObject = JSONObject(otherLocations).apply {
+        // Some versions of org.json's JSONObject(Map) store the reference directly
+        // without copying, so the map must be mutable. And even on other versions,
+        // R8 optimizations can make things go like this.
+        override fun toJSON(): JSONObject = JSONObject(otherLocations.toMutableMap()).apply {
             putIfNotEmpty("fragments", fragments)
             put("progression", progression)
             put("position", position)
@@ -285,7 +288,7 @@ public data class LocatorCollection(
          */
         val title: String? get() = localizedTitle?.string
 
-        override fun toJSON(): JSONObject = JSONObject(otherMetadata).apply {
+        override fun toJSON(): JSONObject = JSONObject(otherMetadata.toMutableMap()).apply {
             putIfNotEmpty("title", localizedTitle)
             putOpt("numberOfItems", numberOfItems)
         }