+++ readium/shared/src/main/java/org/readium/r2/shared/publication/epub/Presentation.kt
@@ -7,6 +7,8 @@
  * LICENSE file present in the project repository where this source code is maintained.
  */
 
+@file:Suppress("DEPRECATION")
+
 package org.readium.r2.shared.publication.epub
 
 import org.readium.r2.shared.publication.Link
@@ -16,6 +18,7 @@ import org.readium.r2.shared.publication.presentation.Presentation
  * Get the layout of the given resource in this publication.
  * Falls back on REFLOWABLE.
  */
+@Deprecated("This was removed from RWPM. You can still use the EPUB extensibility to access the original values.")
 public fun Presentation.layoutOf(link: Link): EpubLayout {
     return link.properties.layout
         ?: layout