+++ readium/shared/src/main/java/org/readium/r2/shared/publication/presentation/Metadata.kt
@@ -6,6 +6,9 @@
  * Use of this source code is governed by a BSD-style license which is detailed in the
  * LICENSE file present in the project repository where this source code is maintained.
  */
+
+@file:Suppress("DEPRECATION")
+
 package org.readium.r2.shared.publication.presentation
 
 import org.json.JSONObject
@@ -13,6 +16,7 @@ import org.readium.r2.shared.publication.Metadata
 
 // Presentation extensions for [Metadata]
 
+@Deprecated("This was removed from RWPM. You can still use the EPUB extensibility to access the original values.", level = DeprecationLevel.ERROR)
 public val Metadata.presentation: Presentation
     get() = Presentation.fromJSON(
         (this["presentation"] as? Map<*, *>)