+++ CHANGELOG.md
@@ -25,6 +25,10 @@ All notable changes to this project will be documented in this file. Take a look
 
 * Removed title inference based on folder names within image and audio archives. Use the archive's filename instead.
 
+#### Shared
+
+* The Presentation Hints properties are deprecated from the Readium Web Publication Manifest models. [See the official documentation](https://readium.org/webpub-manifest/profiles/epub.html#appendix-b---deprecated-properties).
+
 ### Fixed
 
 #### Shared